Fixes for windows warnings savannah task #15808
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@38202 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: cce2b40ec363ec22f3403da881ed1ee4b0269a58

+++ Rec/Tr/TrackProjectors/src/TrajOTCosmicsProjector.cpp
@@ -67,7 +67,7 @@ StatusCode TrajOTCosmicsProjector::project( const LHCb::StateVector& statevector
     // qop-entry of the statevector to store the phase.
     double eventt0 = statevector.parameters()[4] ;
     // ugly const-cast to update the measurement's time-of-flight
-    (const_cast< LHCb::OTMeasurement&>(meas)).setDeltaTimeOfFlight( tof + eventt0 ) ;
+    (const_cast< LHCb::OTMeasurement&>(meas)).setDeltaTimeOfFlight( (float) (tof + eventt0) ) ;
     
     // need to test ambiguity before calling projector!-(
     // call the standard projector (which uses the time-of-flight)
